requirements.txt
python-dotenv==1.2.1
fastapi==0.121.1
uvicorn==0.38.0
google-cloud-firestore==2.21.0
email-validator==2.3.0
PyJWT==2.8.0
websockets-15.0.1